Separated HyperLogLog container implementation into hyperloglog.rs.
